+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - [INTERNAL] Move `super` and `undef` nodes into the methods file.
 - [INTERNAL] Move assign nodes into their own file.
 - [INTERNAL] Move constant nodes into their own file.
+- [INTERNAL] Move `number_arg` nodes into the blocks file.
 
 ## [0.12.2] - 2019-04-30